Move personal access token to CloudTAK
